+++ lib/src/base/platform_streams.dart
@@ -36,6 +36,7 @@ class PlatformStreams {
         StreamController<PlayerIdentifier<void>>.broadcast();
     _recordingAmplitudeController = StreamController<double>.broadcast();
     _recordedBytesController = StreamController<Uint8List>.broadcast();
+    _recordedDurationController = StreamController<Duration>.broadcast();
     await AudioWaveformsInterface.instance.setMethodCallHandler();
   }
 
@@ -55,7 +56,10 @@ class PlatformStreams {
       _completionController.stream;
 
   Stream<double> get onAmplitude => _recordingAmplitudeController.stream;
+
   Stream<Uint8List> get onRecordedBytes => _recordedBytesController.stream;
+
+  Stream<Duration> get onCurrentDuration => _recordedDurationController.stream;
 
   late StreamController<PlayerIdentifier<int>> _currentDurationController;
   late StreamController<PlayerIdentifier<PlayerState>> _playerStateController;
@@ -65,6 +69,7 @@ class PlatformStreams {
   late StreamController<PlayerIdentifier<void>> _completionController;
   late StreamController<double> _recordingAmplitudeController;
   late StreamController<Uint8List> _recordedBytesController;
+  late StreamController<Duration> _recordedDurationController;
 
   void addCurrentDurationEvent(PlayerIdentifier<int> playerIdentifier) {
     if (!_currentDurationController.isClosed) {
@@ -108,6 +113,11 @@ class PlatformStreams {
     _recordedBytesController.add(event);
   }
 
+  void addRecordedDurationEvent(int milliSeconds) {
+    if (_recordedDurationController.isClosed) return;
+    _recordedDurationController.add(Duration(milliseconds: milliSeconds));
+  }
+
   void dispose() {
     _currentDurationController.close();
     _playerStateController.close();
@@ -115,6 +125,7 @@ class PlatformStreams {
     _completionController.close();
     _recordingAmplitudeController.close();
     _recordedBytesController.close();
+    _recordedDurationController.close();
     AudioWaveformsInterface.instance.removeMethodCallHandler();
     isInitialised = false;
   }